requirements.txt
django = 6.0.2
djangorestframework = 3.16.1

pip install -r requirements.txt